+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/TemporaryStateService.cs
@@ -13,6 +13,12 @@ public interface ITemporaryStateService {
     bool IsIacScanningEnabled { get; }
     bool IsSastScanningEnabled { get; }
     bool IsAiLargeLanguageModelEnabled { get; }
+    
+    bool IsTreeViewFilterByCriticalSeverityEnabled { get; set; }
+    bool IsTreeViewFilterByHighSeverityEnabled { get; set; }
+    bool IsTreeViewFilterByMediumSeverityEnabled { get; set; }
+    bool IsTreeViewFilterByLowSeverityEnabled { get; set; }
+    bool IsTreeViewFilterByInfoSeverityEnabled { get; set; }
 }
 
 public class TemporaryStateService : ITemporaryStateService {
@@ -36,6 +42,12 @@ public class TemporaryStateService : ITemporaryStateService {
     public bool IsIacScanningEnabled => _cliStatus?.SupportedModules?.IacScanning == true;
     public bool IsSastScanningEnabled => _cliStatus?.SupportedModules?.SastScanning == true;
     public bool IsAiLargeLanguageModelEnabled => _cliStatus?.SupportedModules?.AiLargeLanguageModel == true;
+
+    public bool IsTreeViewFilterByCriticalSeverityEnabled { get; set; } = false;
+    public bool IsTreeViewFilterByHighSeverityEnabled { get; set; } = false;
+    public bool IsTreeViewFilterByMediumSeverityEnabled { get; set; } = false;
+    public bool IsTreeViewFilterByLowSeverityEnabled { get; set; } = false;
+    public bool IsTreeViewFilterByInfoSeverityEnabled { get; set; } = false;
 
     public TemporaryStateService(ILoggerService loggerService) {
         _loggerService = loggerService;