+++ src/extension/Cycode.VisualStudio.Extension.Shared/VSCommandTable.cs
@@ -23,6 +23,12 @@ internal sealed class PackageIds {
     public const int TreeViewExpandAllCommand = 0x1057;
     public const int TreeViewCollapseAllCommand = 0x1058;
 
+    public const int TreeViewFilterByCriticalSeverityCommand = 0x1059;
+    public const int TreeViewFilterByHighSeverityCommand = 0x1060;
+    public const int TreeViewFilterByMediumSeverityCommand = 0x1061;
+    public const int TreeViewFilterByLowSeverityCommand = 0x1062;
+    public const int TreeViewFilterByInfoSeverityCommand = 0x1063;
+
     public const int TopMenuCycodeCommand = 0x1103;
     public const int TopMenuOpenSettingsCommand = 0x1104;
 }